+++ client/config/config.go
@@ -223,7 +223,7 @@ type Config struct {
 	StateDBFactory state.NewStateDBFunc
 
 	// HostVolumes is the set of configured host volumes
-	HostVolumes map[string]*config.HostVolumeConfig
+	HostVolumes map[string]*structs.HostVolumeConfig
 }
 
 func (c *Config) Copy() *Config {
